[SKINCARE-1] revamping and succeed for etl sephora done
